requirements.txt
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
Flask-SocketIO==5.3.2
mysql-connector-python==8.0.33
pydicom==2.4.1
numpy==1.24.3
tensorflow==2.12.0
torch==2.0.1
torchvision==0.15.2
Pillow==9.5.0
